Invoke-CMApplyDriverPackage.ps1 version 2.0.6
Removed delimiter from the $Time join in Write-CMLogEntry function. This fixes CMTrace formatting.
Moved some of the 'Attempting' logging before the step it logs rather than after it.

Support web service high availability:
Changed URI parameter to a string array which is backwards compatible with a single string.  Multiple URIs can be specified by delimiting them with commas.
Added Servers parameter which takes the FQDN of one or more servers running the web service.  Multiple servers can be specified by delimiting them with commas.
Made URI and Server parameters optional. If neither is specified then assume the web service is running on management points. The MP list will be grabbed from either WMI or the MP the task sequence is using.
When multiple URIs and/or servers are used then randomize the order that servers are contacted. This isn't true load balancing but it's close enough.