Update Invoke-CMDownloadBIOSPackage.ps1
Change conditional operator when looking for package matches against a computer model. This is because Dell has model names like 'OptiPlex 9020' and 'OptiPlex 9020 AIO' which will both match if BIOS packages have been downloaded for both models and a computer that is an OptiPlex 9020 runs this script. By checking against the PackageName and using a wildcard at the beginning with a like operator instead of a match, this should help an OptiPlex 9020 not return a result for the 9020 AIO.