Update Invoke-CMApplyDriverPackage.ps1
Inside Invoke-CMDownloadContent function, changed conditional to check for _SMSTSInWinPE TS variable rather than the existence of c:\windows\ccm\osddownloadcontent.exe. The latter path can exist even when inside WinPE when reimaging a computer that already had Windows/CCM client on it, and can cause the TS to fail when DLLs for the OSDDownloadContent executable can't be properly loaded. Due to this error, I was getting "package content download process failed 1073741511" as a return error from the Invoke-CMDownloadContent function.