Update Invoke-CMApplyDriverPackage.ps1
If the apply OS step is done in a child task sequence, OS Image information can't be retrieved because the SMSTSPackageID of the main Task Sequence is used by default. Added an optional parameter ChildTSVariableName that accepts a task sequence variable which will allow the SMSTSPackageID of the child task sequence that has the apply OS step to be specified.