Fix for accessing content directly from DP
Fixes #14 and #41 This adds an optional PkgShare parameter to use when calling Invoke-CMDownloadContent.ps1 that allows the driver package to be downloaded from the package share using the Copy-Item cmdlet rather than OSDDownloadContent.exe.  It's been tested to work in my environment on single package matches.  I haven't attempted a multiple package match scenario with this.

The caveat here is that that the package share name will need to be the same on all DPs for this to work.  I suppose setting a custom package share name for all driver packages would suffice here.  In my case, the site server I'm turning up has a default content share of SMSPKGF$, but the DPs I'm going to start turning up will most likely have SMSPKGD$.

Hopefully somebody else in the community is better at scripting than myself and could propose changes to this to determine the package share name containing the driver packages and render the PkgShare parameter unnecessary.